refactor: 변수, 함수 성격에 맞게 폴더링
