+++ src/basic/lib/store.js
@@ -1,0 +1,6 @@
+export const store = {
+  lastSelectedId: null,
+  bonusPts: 0,
+  totalAmt: 0,
+  itemCnt: 0,
+};
